Only show pie chart/portfolio on overview tab
